+++ simplesocialmediawebapplication/src/main/java/com/zerplabsintern/simplesocialmediawebapplication/service/impl/PostImagesServiceImpl.java
@@ -4,6 +4,7 @@ import java.util.Base64;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import com.zerplabsintern.simplesocialmediawebapplication.dto.PostImagesDto;
@@ -12,6 +13,7 @@ import com.zerplabsintern.simplesocialmediawebapplication.exception.PostImagesSe
 import com.zerplabsintern.simplesocialmediawebapplication.exception.UserServiceException;
 import com.zerplabsintern.simplesocialmediawebapplication.repository.PostImagesRepository;
 import com.zerplabsintern.simplesocialmediawebapplication.repository.PostRepository;
+import com.zerplabsintern.simplesocialmediawebapplication.repository.UserRepository;
 import com.zerplabsintern.simplesocialmediawebapplication.service.CompressImageService;
 import com.zerplabsintern.simplesocialmediawebapplication.service.PostImagesService;
 
@@ -27,8 +29,11 @@ public class PostImagesServiceImpl implements PostImagesService {
     @Autowired
     private CompressImageService compressImageService;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @Override
-    public PostImagesDto addPostImages(PostImagesDto postImagesDto) {
+    public PostImagesDto addPostImages(PostImagesDto postImagesDto, UserDetails currentUser) {
         try {
 
             PostImages postImages = new PostImages();
@@ -53,7 +58,14 @@ public class PostImagesServiceImpl implements PostImagesService {
 
             if(postImagesDto.getPostId() != null ) {
 
-                postImages.setpIPost(postRepository.findById(postImagesDto.getPostId()).get());
+                if (postRepository.findById(postImagesDto.getPostId()).get().getpUser().getId() == userRepository.findIdbyemailId(currentUser.getUsername()) ) {
+
+                    postImages.setpIPost(postRepository.findById(postImagesDto.getPostId()).get());
+                }
+                else {
+                    throw new PostImagesServiceException("the given post id user is not matching with the logged in user.");
+                }
+
             }
             else {
                 throw new PostImagesServiceException("postId field cannot be null");
@@ -69,16 +81,24 @@ public class PostImagesServiceImpl implements PostImagesService {
     }
 
     @Override
-    public boolean deletePostImages( Long id ) {
+    public boolean deletePostImages( Long id, UserDetails currentUser ) {
         try {
-            if(postImagesRepository.findById(id).isPresent()){
 
-                postImagesRepository.deleteById(id);;
-                return true;
+            if(userRepository.findIdbyemailId(currentUser.getUsername()) == postRepository.findById(postImagesRepository.findById(id).get().getpIPost().getId()).get().getpUser().getId() ) {
+
+                if(postImagesRepository.findById(id).isPresent()){
+    
+                    postImagesRepository.deleteById(id);;
+                    return true;
+                }
+                else {
+                    throw new PostImagesServiceException("the id given does not correspond to any of the post images, so please check the id and send the request again ");
+                }
             }
             else {
-                throw new PostImagesServiceException("the id given does not correspond to any of the post images, so please check the id and send the request again ");
+                throw new PostImagesServiceException("the user corresponding to this request does not match with the logged in user");
             }
+
         } catch (Exception e) {
             throw new PostImagesServiceException("there is a exception while trying to delete the postImages... ");
         }
